Add mobile pick-and-place-with-navigation FSM expert
Introduce a continuous mobile manipulation expert that navigates to a
pickup object, picks it, navigates to a receptacle, and places it, reusing
the existing A*/PurePursuit nav policy and the pick-and-place manipulation
planner via composition.

New:
- configs/mobile_pick_and_place_config.py: MobileFrankaPickAndPlaceConfig
  (+ task/sampler/policy sub-configs) on the mobile Franka robot.
- tasks/mobile_pick_and_place_task.py: MobilePickAndPlaceTask, inheriting the
  PickAndPlaceTask success predicate and adding the nav-task interface.
- tasks/mobile_pick_and_place_task_sampler.py: MobilePickAndPlaceTaskSampler,
  reusing pnp object/receptacle selection + nav-style navigable robot placement.
- policy/solvers/mobile_pick_and_place_state_machine_policy.py: FSM policy
  (NAV_TO_OBJ -> PICK -> NAV_TO_RECEPTACLE -> PLACE -> DONE). Manipulation
  primitives are built AFTER navigation parks the base, with base-locked IK and
  adaptive pregrasp/preplace stand-offs to cope with near-reach-limit parking.

Fixes:
- kinematics/parallel/warp_kinematics.py: pass explicit dtype to wp.from_numpy
  for the jacobian mask and qpos copies; warp 1.15 otherwise infers a vector
  dtype and warmup_ik raises "Incompatible array data types" (broke all
  parallel-IK manipulation warmup, not just the mobile robot).
- utils/grasp_sample.py: register "mobile_pick_and_place" as a pnp task type
  for grasp loading.

Verified end-to-end: judge_success()==True on house_id=2 seed=1.

Copilot-Session: 80c883b1-7903-439e-822a-7ff79156d44e

+++ molmo_spaces/kinematics/parallel/warp_kinematics.py
@@ -605,11 +605,14 @@ class SimpleWarpKinematics(ParallelKinematics):
             ik_args.dt.fill_(dt)
             wp.copy(
                 ik_args.jacobian_mask,
-                wp.from_numpy(self._create_jacobian_mask(batch_size, unlocked_move_group_ids)),
+                wp.from_numpy(
+                    self._create_jacobian_mask(batch_size, unlocked_move_group_ids),
+                    dtype=wp.int32,
+                ),
             )
 
             q0_arr = self._dicts_to_qpos_arr(q0_dicts)
-            wp.copy(data.qpos, wp.from_numpy(q0_arr))
+            wp.copy(data.qpos, wp.from_numpy(q0_arr, dtype=wp.float32))
 
             for i in range(max_iter):
                 if self._device.startswith("cuda"):
